requirements.txt
python-dotenv
qdrant-client[fastembed]
openai
mistralai
pymupdf
python-pptx
python-docx
psycopg2-binary
pillow
pydantic
python-multipart
rich
pyfiglet
google-genai
pdfplumberlangchain-ollama
langchain-ollama
